Make libaidlcommonsupport available to apex modules
Making libaidlcommonsupport available to apex has the following impact
on affected apex modules. All sizes were collected on top-of-tree
git_master for cf_x86_64_phone-userdebug.

libaidlcommonsupport.a:
* lib: 93K
* lib64: 128K

Diff from minimally including libaidlcommonsupport:
* lib/libneuralnetworks.so: +1K
* lib/libneuralnetworks_shim.so: +0K
* lib64/libneuralnetworks.so: +1K
* lib64/libneuralnetworks_shim.so: +0K

There's only a small change in size to libneuralnetworks.so and
libneuralnetworks_shim.so because both had code that performed the same
conversions that are done in libaidlcommonsupport.

Bug: 183118727
Test: mma
Change-Id: I04403eab718161d6a77499e2a9af82ee61ed42af