Add build rules for updated SDK 30 android.jars
These build rules are work-in-progress libraries to produce SDK jars
based on SDK 30 but with updated mainline module stubs.

They aren't yet ready to be used, as some of the components need
tweaking (for example, the android-non-updatable prebuilt does not
have resources in it)

Migrate in the android.jar defaults from frameworks/base to avoid taking
a dependency on that repository.

Bug: 184839225
Test: m android_30_modules_current{,_system}
Change-Id: Ie696406c57f437c9c46d727311ff2956b2c8c4ea