Add SettingsLib libraries used by PermissionController to allowed_deps.
We need to use these SettingsLib preferences in order to adapt to the
new Settings theme in S+.

The minSdkVersion:21 here is technically incorrect and should be 30
instead. This will be addressed in a future change that fixes it for all
SettingsLib libraries and cleans up the API Lint baseline text files as
well.

Bug: 14490549
Test: presubmit
Change-Id: Id19c89116251ab4e7ce23ec1f80e272d0d8ccb13