Add car-rotary-lib to allowed dependencies list
The new library is extracted from the existing car-ui-lib. It contains
rotary specific code, which is already used by apps supporting rotary.

> How big is the binary size increase for affected APEXes?
It doesn't increase the size. In fact, it reduces the size if an app
wants to support rotary but doesn't want to use the whole car-ui-lib.

> Is the new dependency committed to support previous platform releases?
Yes.

> Is the new dependency being developed AOSP-first or internal?
AOSP-first.

> What’s the testing strategy for the new dependency?
> Does it have its own tests, and are you adding integration tests?
It has its own tests, such as FocusParkingViewTest.

Bug: 188915160
Test: make
Change-Id: Ie7b10416f845c35a19309c5091a394d65c0efe40