Update allowed_deps for bumped NetworkStack
Add NetworkStackApi32Shims to allowed_deps for NetworkStack.
As documented on the build rule "API 32" is not technically correct and
should be "in-progress API", but that naming has always avoided renames
at SDK finalization time.

The new target is used in development NetworkStack builds that are not
pushed to released devices (NetworkStackNext) to allow it to use
non-finalized APIs.

Ignore-AOSP-First: cherry-pick prepared separately
Bug: 189581241
Test: m
Change-Id: Ie274b4a1ea66bbbe26c7d7a78c90a19a746ea84e