Introduce build script for mainline modules.
The build_unbundled_mainline_module.sh builds for a given product
a list of mainline modules. In AOSP it defaults to building all
AOSP modules.

Individual mainline_modules_<arch>.sh scripts are meant to build per
product targets for individual invocation on build servers. For example,
on AOSP presubmit we would want to build mainline_modules_x86_64.

The build script is meant to be re-used from internal version of the
scripts that would pass com.google variant of the modules.

Note that the script excludes bundles and stubs, since those are not
need to be built in AOSP. There will be a separate script internally
for those.

Bug: 191965323
Test: call the scripts, inspect out/dist/ directory.
Change-Id: I13dc723dd26fd8a2a43f46c3186513bcf5519fb5