Add commit hook to require extra info in allowed_deps.txt additions.
For any CL that adds a new dependency to allowed_deps.txt, this will
require that extra information is provided in the commit message:
  Apex-Size-Increase:
  Previous-Platform-Support:
  Aosp-First:
  Test-Info:

Any CL that does not modify allowed_deps.txt, or that doesn't add new
dependency names to it, will be ignored.

Also open up ownership of allowed_deps.txt to allow anyone to make
changes there. We will use the information provided in the commit
message to verify additions periodically, rather than doing this in
code review.

Test: for sha in $(git log --format=%H build/allowed_deps.txt);
       do ./tools/check_allowed_deps.py $sha;
      done
Bug: 195303213
Change-Id: Ief9125d1eb51b29a7d3f05e5485ff41841d1d494