Add glide prebuilt dependencies to allowed_deps
Apex-Size-Increase: 4292 bytes
Previous-Platform-Support: Yes
Aosp-First: Glide dependency is AOSP-first, using the dependency to develop an
internal project.
Ignore-AOSP-First: There are changes in topic which cannot be made in AOSP. The updating of
glide dependency in AOSP would be done by a separate change.
Test-Info: We will be adding integration and manual tests for the image loading
experience.
Test: m
Bug: 195008673

Change-Id: I35641941b972ab87fe4091c36a6fded1c2c54705