Enhance documentation in mainline stubs
Metalava has a feature to "enhance documentation" which is mostly based
on auto-generating documentation text based on source-retention
annotations. This is turned on automatically for doc stubs, but here
we are enabling it for module's jar stubs as well.

The point of this is to be able to replace the module source code with
module jar stub source code when generating doc stubs -- something
previously not possible because most source-retention annotations are
not present in the jar stubs.

In order for the enhancement to work properly, metalava needs to be
passed the annotation source code rather than the annotations library.
This is because of various javadoc tags controlling what documentation
is generated, e.g. @paramDoc, @returnDoc etc.

Bug: 186197911
Test: manually inspect module stub source code is enhanced
Change-Id: I3fda4cd64e49c1da33fe614c30aafd9fa0dba3b9