Add STANDALONE_SYSTEMSERVER_JARS to classpath.proto.
This is an environment variable that lists the jars that system server
loads dynamically using separate classloaders. The list is derived from
PRODUCT_STANDALONE_SYSTEM_SERVER_JARS and
PRODUCT_APEX_PRODUCT_STANDALONE_SYSTEM_SERVER_JARS in Makefile. It will
be comsumed by odrefresh to determine what jars to preopt on early boot.

Note that the variable should not end with "CLASSPATH" because the list
is not used by runtime as a classpath. It is just a colon-separated list
of jars. System server loads the jars separately with paths hardcoded in
the code.

Bug: 203198541
Test: manual - 1. Build an image.
  2. Flash the image to a device.
  3. adb shell echo \$STANDALONE_SYSTEMSERVER_JARS
  4. See the correct list of jars.
Change-Id: I5a9653c606969aaa8083e8c8daa9d48b99663790