Update allowed_deps for AppSearch apex.
icing-java-proto-lite contains protobuf messages for working with icing
lib.

icing lib is the library underlying AppSearch and we don't expect any
other module to be interested in this lib or these protos.

The protos and the operation of icing lib are verified through
AppSearch's CTS and MTS tests.

Aosp-First: internal
Test-Info: Tested as part of CtsAppSearchTestCases
Previous-Platform-Support: Yes. The dependency is statically compiled into AppSearch and will support all platforms that AppSearch supports.
Apex-Size-Increase: 448KiB
Bug: 146218515
Test: Builds; passes CTS tests
Change-Id: I74caed11a00e99a48be83628768a1a5cae3ac2aa