Make first_release required
(cherry picked from commit 33e939944570534d7e182a5a6687b669410583c1)

Defaulting the first_release to LATEST is dangerous because it is easy
to forget and any module that does not provide it will not be built for
a release specific build.

Adds Tiramisu for the mainline modules that were missing it and
defaults it to LATEST for BundledMainlineModules as they don't need to
be built for a previous release.

Bug: 228807460
Test: packages/modules/common/build/mainline_modules_sdks.py
Merged-In: I151773f80e37a6dd7de7b6df9cfa3f0a624d38f4
Change-Id: I151773f80e37a6dd7de7b6df9cfa3f0a624d38f4