Code refactoring to build targets in a separate method.
The build_snapshot method is creating a list of targets and building
them by calling soong.
The code sets some environment variables and builds the targets. This
code can be made reusable by putting it into a separate method and
passing the list of targets to be built. This will avoid code
redundancy.

Bug: 230609867
Test: The code builds on the intellij and successfully builds all the
targets passed to this newly added method.

Change-Id: I22566ca438b2da0cf832416466f4e3affb1c976e