Allow permission to be treated as optional at build time
Although Permission is not an optional module on GMS capable devices it
is optional for non-GMS capable devices and so needs to support being
treated as optional at build time. Which means that an OEM needs to be
able to choose whether to use the prebuilt Permission module or build
their own module from source.

This is safe even though permission is not an optional mainline
module because devices that support mainline must pass GTS and
testModulesSignedCorrectly and testModulesPreloadedCorrectly in GTS
currently tests this, along with testMainlineApprovedTrain which checks
mandatory/optional, signing keys, and if what's preloaded are part of
approved trains.

Bug: 238175656
Test: packages/modules/common/build/mainline_modules_sdks.sh
      # Check that permission sdk snapshot uses a permission specific
      # Soong config variable.
Change-Id: I8278622fedbba0292f7cab1114170394cf429849