Used fixed timestamps for the zip contents
Previously, while the snapshot zip files generated by Soong had entries
with fixed timestamps the zip files produced by this script did not
which meant that the contents of the zip files depended on when the zip
file was created. That meant that in order to compare a change affected
the zip file contents it was necessary to unpack the before and after
zip files in order to compare them.

This change causes all zip files created by the script to use the same
fixed timestamp for each of its entries as soong_zip and ziptime do so
it is now possible to check that the contents of sdk snapshot zip files
are the same without unpacking them.

Bug: 232401814
Test: BUILD_NUMBER=fixed packages/modules/common/build/mainline_modules_sdks.sh
      # Remove api diff files as they contain file stamps of generated files so
      # differ every time they are generated.
      find out/dist/mainline-sdks -name \*txt | xargs rm
      # Save the snapshots away.
      mv out/dist/mainline-sdks before-changes
      # Repeat the first two steps above and then run the following to verify
      # that simply running the script twice results in different sdk contents.
      meld before-changes out/dist/mainline-sdks
      # Apply this change and then run the above again but this time the sdks
      # will be identical.
Change-Id: I28bf78783c827abffa6b715c62ca5948b3ee19b5