Deal with inconsistent module names
Conscrypt and ART have some history of using dist_stem names in
prebuilts/sdk that are different from the module names of their
java_sdk_library, leading to same naming inconsistensies.

Update the finalize script to reconcile these inconsistensies during
import.

Note that the build has already been updated to throw an error when
inconsistensies are discovered, so this update to the script is to
make sure the next import of conscrypt/art goes smoothly.

Test: finalize_sdk.py with conscrypt/art
Change-Id: If82fadf73d508d4807aef11b960e0860d54874b7