Add non-updatable-framework-module-defaults
These java_defaults are useful for the java_sdk_libraries that belong to
non-updatable APEXes.

Defaults are very similar to the framework-module-defaults, the main
differences are:

1. non-updatable-framework-module-defaults enables @TestApi surface
2. non-updatable-framework-module-defaults doesn't enable
   strict_updatability_linting

Allow list of the users of these new defaults is controlled by the
defaults_visibility, so far only Virtualization apex is allowed to use
non-updatable-framework-module-defaults.

Bug: 243512044
Test: m
Change-Id: Id06fb6fc43be78c1bd92c29e7593df89e412f598