Use libbitflags-1.3.2
external/rust/bitflags has been upgraded to version 2.2.1. However, some
external crates still depend on the previous version (1.3.2). We are keeping
1.3.2 until all crates have transitioned to a more recent version.

Test: Build
No-Allowed-Deps-Check: Not a new dependency, just an older version of an
already existing dependecy being kept back after an update.
Ignore-AOSP-First: This is a cherry-pick from AOSP because changes to
Rust crates are not automerged from AOSP.

Change-Id: Iee719e1435c0b36b1b51624a63ffe6eab4f8f049