ApkInApex: set min_sdk_version of test APEX
The ApkInApex test installs a test APEX and inspects what was and what
wasn't successfully installed. However, the APEX blueprint does not
explicitly set its min_sdk_version, so the build will assign it a
defult value, which currently happens to be 34. This means that
ApkInApex, which should be possible to run on SDK 33+, fails to install
its APEX on SDK 33.

Explicitly set the min_sdk_version of the test APEX to 33.

Note: with this CL, ApkInApexTest no longer crashes. It does however
fail. That failure will be fixed in a follow-up CL.

Ignore-AOSP-First: test developed internally
Test: atest ApkInApexTest # test will fail, but will at least not crash
Bug: 289877353
Change-Id: Iab7f7d4f6eefca0873a15360477146022cb5422f