UpdatableSharedLibsTest: set min_sdk_version of test APEX
The UpdatableSharedLibsTest test installs a test APEX as part of the
test. However, the APEX blueprint does not explicitly set its
min_sdk_version, so the build will assign it a defult value, which
currently happens to be 34. This means that UpdatableSharedLibsTest will
push an incompatible APEX to T devices, which will result in a bootloop
when the test reboots the device.

Test: atest UpdatableSharedLibsTest
Bug: 289877353
Change-Id: Iab7f7d4f6eefca0873a15360477146022cb5422f