ApkInApexTest: set min_sdk_version of all helper apps
Explicitly set the min_sdk_version of all apps used as part of the
ApkInApexTest test. If unset, they will default to the current SDK, but
the tests are intended to be run on SDK 33 devices, and a higher
min_sdk_version will prevent installs on SDK 33.

Ignore-AOSP-First: ApkInApexTest not available in AOSP yet
Bug: 289877353
Bug: 293264320
Test: atest ApkInApexTest
Change-Id: Id93dfcdd5a5272270ae7b8d53b29cc5b42943931