Add an option to skip the allowed deps check.
This is for use on thin manifests where there are `apex` Soong modules
in the tree that lack dependencies to be built. Since the APEX
singleton in Soong adds all such modules to the apex-allowed-deps-check
target it won't work, even if SOONG_ALLOW_MISSING_DEPENDENCIES is used.

Also use the apex-allowed-deps-check phony target rather than
referencing the check file directly.

Test: env TARGET_BUILD_VARIANT=userdebug \
        TARGET_BUILD_APPS="com.android.art com.android.art.debug" \
        BUILD_BROKEN_DISABLE_BAZEL=true \
        packages/modules/common/build/mainline_modules_sdks.sh \
        --skip-allowed-deps-check
  on master-art with https://r.android.com/2743421, which adds a
  project with the com.andoid.tethering APEX, which cannot be built on
  master-art.
Test: env TARGET_BUILD_VARIANT=userdebug \
        TARGET_BUILD_APPS="com.android.art com.android.art.debug" \
        packages/modules/common/build/mainline_modules_sdks.sh \
        --build-release latest
  on AOSP main with libbase removed from allowed_deps.txt, and verify
  that it fails.
Bug: 300072791
Change-Id: I313905e1f00614b263d8240d01504805380e3076