Add an allowed_deps.txt dependency
Add an allowed_deps.txt dependency for a module that isn't currently
used.

An existing library has been split into two.

Bug: 312242276
Test: atest affected tests
Apex-Size-Increase: 0 because it's a split of an existing library
Previous-Platform-Support: N/A - apex is not deployed
Ignore-Aosp-First: Internal change - no need for AOSP because not deployed. May CP
Test-Info: Split of existing code, existing tests continue to exist
Change-Id: I744c19dd4d398a4a8bbcc45b205fb4030527ce93