Add next BuildRelease for use by SDK finalization
Currently, the SDK finalization script consumes the SDK snapshots from
the `for-latest-build` directory. However, in trunk stable that is no
longer viable as that directory will include flagged APIs but the
finalized SDK cannot use flagged APIs. It is not possible to remove
flagged APIs from `for-latest-build` as the snapshots in there are
dropped into Android itself and so must include flagged APIs.

This change adds a new `NEXT` `BuildRelease` which will create a
`for-next-build` that will contain SDK snapshots that do not contain
flagged APIs. It also adds an `include_flagged_apis` property which
is `False` for everything except `LATEST`. For now that property is
for informational purposes but a follow up change will use it to
determine whether the SDK snapshots include flagged APIs.

A separate follow up change will modify the finalize script to use
the `for-next-build` directory. It is done separately as it will
take a while for this change to have an impact on the build targets
that the finalize script consumes.

Bug: 313065235
Test: packages/modules/common/build/mainline_modules_sdks.sh
      # Make sure that out/dist/mainline-sdks/for-next-build is
      # identical to for-latest-build.
Change-Id: Ib6d18fda129d081e8bbc8bf148cb5d7f38c070ba