Finalize using the for-next-build SDK snapshots
Previously, the SDK finalization script consumed the SDK snapshots from
the `for-latest-build` directory. However, in trunk stable that is no
longer viable as that directory will include flagged APIs but the
finalized SDK cannot use flagged APIs. It is not possible to remove
flagged APIs from `for-latest-build` as the snapshots in there are
dropped into Android itself and so must include flagged APIs.

This change switches the script to use `for-next-build` which is
currently identical to `for-latest-build` but will, eventually contain
no flagged APIs.

Bug: 313065235
Test: TBD
Change-Id: I2e7a63a02aa218e259d6b970d64016d15f42fb0f