finalize_sdk.py: add (optional) option to set build target
finalize_sdk.py downloads artifacts from an ab/ build and implements
heuristics to determine which of the build's targets to download.

With the introduction of mainline monthly missions, we can no longer
download artifacts from the git_main builds -- these lack targets for
the current mainline_<year>_<month> release config. Instead, we have to
fetch artifacts from the current mission's release branch
(git_aml-<year>-<month>-release).

The release config part is omitted from build targets on release
branches. For instance, on git_aml-2025-06-release, the target that
would be named mainline_modules_sdks-mainline_2025_06-user on git_main
is named mainline_modules_sdks-user.

This causes the heuristics in finalize_sdk.py to break: not only is the
release config omitted, the target is -user instead of -userdebug.

To combat this, add a new, optional command line option to explicitly
set the build target. To not break update_prebuilts.py, if the new
option is not provided, fall back on the old heuristics.

Bug: N/A
Test: manual
Flag: EXEMPT finalization script
Change-Id: I1bf0ef8413bd1cd628e29a463a555f1370d08fe8